lesson9/data_loader/requirements.txt
aiohttp
aiofiles
load_dotenv